test(GameBoard.test.tsx): Remove test for default board creation, make sure GameBoard always receive
GameBoard now requires board data to be given to it via props, since creating a board with no Game
being created would be unusable. Remove test that checked to see if a board would still be created
even if no data existed, since that should actually throw an error. In test to make sure GameBoard
renders a table, give it some board data to use.